+++ CHANGELOG.md
@@ -5,8 +5,8 @@ FEATURES:
 * add `autoscaling_config` to Data Proc subcluster specification
 * add `ip_address` and `ipv6_address` attributes  to `network_interface` entity in `yandex_compute_instance_group`
 * **New Resource** `yandex_vpc_default_security_group` for network's default security group managment
-
 ENHANCEMENTS:
+* mdb: add `maintenance_window` attribute in `yandex_mdb_mysql_cluster` resource and data source
 * managed-redis: added 'notify_keyspace_events', 'slowlog_log_slower_than', 'slowlog_max_len' and 'databases' fields
 * add `maintenance_window` section in `yandex_mdb_clickhouse_cluster` resource and data source